chore: apply shafu-formatter to all 15 Solidity source files (B1N-95)
- Ran shafu-formatter on all src/**/*.sol files (8 core, 3 interfaces, 4 mocks)
- Changes: uint256→uint shorthand, import alignment, multi-line function sigs,
  variable declaration alignment, double-space before brace
- Fixed shafu bugs: broken compound operators (==, \!=, <=, +=, :=),
  stray { after interface declarations and empty function bodies
- Restored EIP-712 type hash strings to use uint256 (spec requirement)
- Added Makefile with `make fmt` for repeatable formatting
- All 331 tests pass, zero functional changes

+++ src/core/AddressBook.sol
@@ -32,71 +32,101 @@ contract AddressBook is Initializable, UUPSUpgradeable {
     event WhitelistUpdated(address indexed oldAddress, address indexed newAddress);
     event BatchSettlerUpdated(address indexed oldAddress, address indexed newAddress);
 
-    modifier onlyOwner() {
+    modifier onlyOwner()  {
         if (msg.sender != owner) revert OnlyOwner();
         _;
     }
 
     /// @custom:oz-upgrades-unsafe-allow constructor
-    constructor() {
+    constructor()  {
         _disableInitializers();
     }
 
-    function initialize(address _owner) external initializer {
+    function initialize(address _owner)
+        external
+        initializer
+    {
         if (_owner == address(0)) revert InvalidAddress();
-        owner = _owner;
+           owner   = _owner;
     }
 
-    function transferOwnership(address _newOwner) external onlyOwner {
+    function transferOwnership(address _newOwner)
+        external
+        onlyOwner
+    {
         if (_newOwner == address(0)) revert InvalidAddress();
-        pendingOwner = _newOwner;
+           pendingOwner = _newOwner;
         emit OwnershipTransferStarted(owner, _newOwner);
     }
 
-    function acceptOwnership() external {
+    function acceptOwnership()
+        external
+    {
         if (msg.sender != pendingOwner) revert OnlyPendingOwner();
         emit OwnershipTransferred(owner, msg.sender);
-        owner = msg.sender;
+        owner        = msg.sender;
         pendingOwner = address(0);
     }
 
-    function setController(address _controller) external onlyOwner {
+    function setController(address _controller)
+        external
+        onlyOwner
+    {
         if (_controller == address(0)) revert InvalidAddress();
         emit ControllerUpdated(controller, _controller);
         controller = _controller;
     }
 
-    function setMarginPool(address _marginPool) external onlyOwner {
+    function setMarginPool(address _marginPool)
+        external
+        onlyOwner
+    {
         if (_marginPool == address(0)) revert InvalidAddress();
         emit MarginPoolUpdated(marginPool, _marginPool);
         marginPool = _marginPool;
     }
 
-    function setOTokenFactory(address _oTokenFactory) external onlyOwner {
+    function setOTokenFactory(address _oTokenFactory)
+        external
+        onlyOwner
+    {
         if (_oTokenFactory == address(0)) revert InvalidAddress();
         emit OTokenFactoryUpdated(oTokenFactory, _oTokenFactory);
         oTokenFactory = _oTokenFactory;
     }
 
-    function setOracle(address _oracle) external onlyOwner {
+    function setOracle(address _oracle)
+        external
+        onlyOwner
+    {
         if (_oracle == address(0)) revert InvalidAddress();
         emit OracleUpdated(oracle, _oracle);
         oracle = _oracle;
     }
 
-    function setWhitelist(address _whitelist) external onlyOwner {
+    function setWhitelist(address _whitelist)
+        external
+        onlyOwner
+    {
         if (_whitelist == address(0)) revert InvalidAddress();
         emit WhitelistUpdated(whitelist, _whitelist);
         whitelist = _whitelist;
     }
 
-    function setBatchSettler(address _batchSettler) external onlyOwner {
+    function setBatchSettler(address _batchSettler)
+        external
+        onlyOwner
+    {
         if (_batchSettler == address(0)) revert InvalidAddress();
         emit BatchSettlerUpdated(batchSettler, _batchSettler);
         batchSettler = _batchSettler;
     }
 
-    function _authorizeUpgrade(address) internal override onlyOwner {}
+    function _authorizeUpgrade(address)
+        internal
+        override
+        onlyOwner
+        {}
 
-    uint256[42] private __gap;
+    uint[42] private __gap;
 }
