+++ src/core/OTokenFactory.sol
@@ -29,8 +29,8 @@ contract OTokenFactory is Initializable, UUPSUpgradeable {
         address indexed underlying,
         address strikeAsset,
         address collateralAsset,
-        uint256 strikePrice,
-        uint256 expiry,
+        uint strikePrice,
+        uint expiry,
         bool isPut
     );
 
@@ -41,24 +41,27 @@ contract OTokenFactory is Initializable, UUPSUpgradeable {
     error Unauthorized();
 
     /// @custom:oz-upgrades-unsafe-allow constructor
-    constructor() {
+    constructor()  {
         _disableInitializers();
     }
 
-    function initialize(address _addressBook) external initializer {
+    function initialize(address _addressBook)
+        external
+        initializer
+    {
         if (_addressBook == address(0)) revert InvalidAddress();
-        addressBook = AddressBook(_addressBook);
+           addressBook   = AddressBook(_addressBook);
     }
 
     function createOToken(
         address _underlying,
         address _strikeAsset,
         address _collateralAsset,
-        uint256 _strikePrice,
-        uint256 _expiry,
+        uint _strikePrice,
+        uint _expiry,
         bool _isPut
-    ) external returns (address) {
-        if (_underlying == address(0) || _strikeAsset == address(0) || _collateralAsset == address(0)) {
+    ) external returns (address)  {
+        if (_underlying == address(0) || _strikeAsset == address(0) || _collateralAsset == address(0))  {
             revert InvalidAddress();
         }
         if (_strikePrice == 0) revert InvalidStrikePrice();
@@ -79,14 +82,19 @@ contract OTokenFactory is Initializable, UUPSUpgradeable {
         address oTokenAddress = address(oToken);
         oTokens.push(oTokenAddress);
         isOToken[oTokenAddress] = true;
-        getOToken[paramsHash] = oTokenAddress;
+        getOToken[paramsHash]   = oTokenAddress;
 
         emit OTokenCreated(oTokenAddress, _underlying, _strikeAsset, _collateralAsset, _strikePrice, _expiry, _isPut);
 
         return oTokenAddress;
     }
 
-    function getOTokensLength() external view returns (uint256) {
+    function getOTokensLength()
+        external
+        view
+        returns
+        (uint)
+    {
         return oTokens.length;
     }
 
@@ -94,32 +102,35 @@ contract OTokenFactory is Initializable, UUPSUpgradeable {
         address _underlying,
         address _strikeAsset,
         address _collateralAsset,
-        uint256 _strikePrice,
-        uint256 _expiry,
+        uint _strikePrice,
+        uint _expiry,
         bool _isPut
-    ) external view returns (address) {
+    ) external view returns (address)  {
         bytes32 paramsHash = _getParamsHash(_underlying, _strikeAsset, _collateralAsset, _strikePrice, _expiry, _isPut);
 
         bytes32 hash =
             keccak256(abi.encodePacked(bytes1(0xff), address(this), paramsHash, keccak256(type(OToken).creationCode)));
 
-        return address(uint160(uint256(hash)));
+        return address(uint160(uint(hash)));
     }
 
     function _getParamsHash(
         address _underlying,
         address _strikeAsset,
         address _collateralAsset,
-        uint256 _strikePrice,
-        uint256 _expiry,
+        uint _strikePrice,
+        uint _expiry,
         bool _isPut
-    ) internal pure returns (bytes32) {
+    ) internal pure returns (bytes32)  {
         return keccak256(abi.encodePacked(_underlying, _strikeAsset, _collateralAsset, _strikePrice, _expiry, _isPut));
     }
 
-    function _authorizeUpgrade(address) internal override {
+    function _authorizeUpgrade(address)
+        internal
+        override
+    {
         if (msg.sender != addressBook.owner()) revert Unauthorized();
     }
 
-    uint256[46] private __gap;
+    uint[46] private __gap;
 }
