+++ src/interfaces/IFlashLoanSimple.sol
@@ -5,14 +5,20 @@ interface IPool {
     function flashLoanSimple(
         address receiverAddress,
         address asset,
-        uint256 amount,
+        uint amount,
         bytes calldata params,
         uint16 referralCode
     ) external;
 }
 
 interface IFlashLoanSimpleReceiver {
-    function executeOperation(address asset, uint256 amount, uint256 premium, address initiator, bytes calldata params)
+    function executeOperation(
+        address asset,
+        uint amount,
+        uint premium,
+        address initiator,
+        bytes calldata params
+    )
         external
         returns (bool);
 }
