+++ src/interfaces/IMarginVault.sol
@@ -11,7 +11,7 @@ library MarginVault {
     struct Vault {
         address shortOtoken; // The oToken this vault has written (sold)
         address collateralAsset; // The ERC20 used as collateral
-        uint256 shortAmount; // Amount of oTokens minted (8 decimals)
-        uint256 collateralAmount; // Amount of collateral deposited (in collateral's decimals)
+        uint shortAmount; // Amount of oTokens minted (8 decimals)
+        uint collateralAmount; // Amount of collateral deposited (in collateral's decimals)
     }
 }
