+++ src/interfaces/ISwapRouter.sol
@@ -7,10 +7,13 @@ interface ISwapRouter {
         address tokenOut;
         uint24 fee;
         address recipient;
-        uint256 amountOut;
-        uint256 amountInMaximum;
+        uint amountOut;
+        uint amountInMaximum;
         uint160 sqrtPriceLimitX96;
     }
 
-    function exactOutputSingle(ExactOutputSingleParams calldata params) external payable returns (uint256 amountIn);
+    function exactOutputSingle(ExactOutputSingleParams calldata params)
+        external
+        payable
+        returns (uint amountIn);
 }
