+++ src/mocks/MockChainlinkFeed.sol
@@ -9,12 +9,14 @@ pragma solidity 0.8.24;
 contract MockChainlinkFeed {
     int256 public price;
 
-    constructor(int256 _price) {
+    constructor(int256 _price)  {
         require(_price > 0, "MockChainlinkFeed: price must be positive");
         price = _price;
     }
 
-    function setPrice(int256 _price) external {
+    function setPrice(int256 _price)
+        external
+    {
         require(_price > 0, "MockChainlinkFeed: price must be positive");
         price = _price;
     }
@@ -22,7 +24,7 @@ contract MockChainlinkFeed {
     function latestRoundData()
         external
         view
-        returns (uint80 roundId, int256 answer, uint256 startedAt, uint256 updatedAt, uint80 answeredInRound)
+        returns (uint80 roundId, int256 answer, uint startedAt, uint updatedAt, uint80 answeredInRound)
     {
         return (1, price, block.timestamp, block.timestamp, 1);
     }
